native cahed buildrunners for all target platforms
Hitting all 3 target platforms with debug builds targeting both x64 and arm64, just like the upcoming manual build guide.

- The windows and linux build tests the vcpkg configuration and reveals any issues in that install path.
- macOS uses a homebrew approach and only needs to get the binaries from mysql qhich takes about a second.
Add a simple check in the top level cmakefile and add homebrew libraries if we're on apple so the buildbot and people can use natively installed libraries.
- The linux docker already tests a native dpendency path, and most packages can be got from apt-get with the exception of a proper mysql, but that takes but a second to download and install.

all caches are about 1.1gb in size, and with the current 4 native builders that's about 4gb in total so there's room for expansion